Some styling improvements and restructuring
