+++ src/App.tsx
@@ -6,6 +6,12 @@ function App() {
     <>
       <header>
         <h1>Secret Santa</h1>
+        <p>
+          By{' '}
+          <a href="https://brianm.me" rel="noopener noreferrer">
+            Brian Mitchell
+          </a>
+        </p>
       </header>
       <Main />
     </>
